rag-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
neo4j>=5.20.0
sentence-transformers>=3.0.0
httpx>=0.27.0
aiosqlite>=0.20.0
openai>=1.65.0
trafilatura>=1.12.0
beautifulsoup4>=4.13.0
youtube-transcript-api>=0.6.3
yt-dlp>=2024.8.6
pypdf>=4.0.0
pymupdf>=1.24.0
python-docx>=1.0.0
python-multipart>=0.0.9
